move exported functions to identity.go for consistency and remove generate.go
