+++ identity.go
@@ -93,3 +93,62 @@ func newProvisionalIdentity(config config, email string) (provisionalIdentity, e
 
 	return provisionalIdentity, nil
 }
+
+func Create(config Config, userID string) (string, error) {
+	return New(config, userID)
+}
+
+func CreateProvisional(config Config, email string) (string, error) {
+	return NewProvisional(config, email)
+}
+
+func GetPublicIdentity(b64Identity string) (string, error) {
+	return GetPublic(b64Identity)
+}
+
+func New(cfg Config, userID string) (string, error) {
+	conf, err := cfg.fromBase64()
+	if err != nil {
+		return "", err
+	}
+	id, err := newIdentity(conf, userID)
+	if err != nil {
+		return "", err
+	}
+	return Base64JsonEncode(id)
+}
+
+func NewProvisional(config Config, email string) (string, error) {
+	conf, err := config.fromBase64()
+	if err != nil {
+		return "", err
+	}
+	id, err := newProvisionalIdentity(conf, email)
+	if err != nil {
+		return "", err
+	}
+	return Base64JsonEncode(id)
+}
+
+func GetPublic(b64Identity string) (string, error) {
+	type anyPublicIdentity struct {
+		publicIdentity
+		PublicSignatureKey  []byte `json:"public_signature_key,omitempty"`
+		PublicEncryptionKey []byte `json:"public_encryption_key,omitempty"`
+	}
+
+	var (
+		pid anyPublicIdentity
+	)
+
+	err := Base64JsonDecode(b64Identity, &pid)
+	if err != nil {
+		return "", err
+	}
+
+	if pid.Target != "user" && pid.Target != "email" {
+		return "", errors.New("Unsupported identity target")
+	}
+
+	return Base64JsonEncode(pid)
+}
